pyproject.toml
[project]
name = "sw-group-tt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "gunicorn>=23.0.0",
    "djangorestframework-simplejwt>=5.5.0",
    "django-filter>=25.1",
]

[dependency-groups]
dev = [
    "flake8>=7.2.0",
    "flake8-pyproject>=1.2.3",
    "pytest>=8.4.0",
    "pytest-django>=4.11.1",
]

[tool.flake8]
max-line-length = 120
exclude = [
    ".venv",
    "__pycache__",
    ".git"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "SWGroup.settings"
